First runnable version of rss-hydrate
A Python/Flask micro web service that hydrates RSS feeds
from Youtube and GitHub.

For Youtube, it wraps the video URL into an embed so it's
viewable straight from RSS reader.

For GitHub, it fetches the commit's diff and makes that visible
directly in the reader.

Dockerized for deployment to AWS Lightsail